sed/backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-restful==0.3.9
elasticsearch==7.15.0
python-dotenv==0.19.1
logstash-logger==0.2.1
werkzeug==2.0.1
jinja2==3.0.1
itsdangerous==2.0.1
markupsafe==2.0.1
six==1.16.0
python-dateutil==2.8.2

# 多格式数据处理依赖
pandas==1.3.5
openpyxl==3.0.9
pymysql==1.0.2
pymssql==2.2.5
pyodbc==4.0.32

# LLM集成
requests>=2.28.0
openai>=0.27.0

# 数据处理扩展
chardet>=5.0.0
xlrd>=2.0.1